agentQ/requirements.txt
# agentQ/requirements.txt
apache-pulsar-client==3.5.0
fastavro==1.9.4
python-dotenv==1.0.1
PyYAML==6.0.1
pyignite==0.5.0
httpx

# Observability
structlog

# Graph
gremlinpython==3.7.0

# OpenTelemetry
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-requests==0.45b0

# Shared Libraries (install in editable mode)
# -e ./shared/q_vectorstore_client
 
# Shared Libraries
pyyaml
-e ./shared/q_auth_parser
-e ./shared/q_pulse_client
-e ./shared/q_vectorstore_client
-e ./shared/q_knowledgegraph_client
-e ./shared/q_ui_schemas

# For AIOps tools
elasticsearch==8.11.0
kubernetes==28.1.0
